[INJIMOB-956]: Replacing the module and package name to InjiVcRenderer
