+++ kotlin/.idea/gradle.xml
@@ -9,8 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/svgrenderer" />
+            <option value="$PROJECT_DIR$/example-app" />
+            <option value="$PROJECT_DIR$/injivcrenderer" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
