+++ kotlin/example-app/proguard-rules.pro
