+++ kotlin/example-app/src/main/res/mipmap-mdpi/ic_launcher.webp
