+++ kotlin/example-app/src/main/res/values/colors.xml
