+++ kotlin/injivcrenderer/proguard-rules.pro
