+++ kotlin/injivcrenderer/src/main/AndroidManifest.xml
