+++ kotlin/injivcrenderer/src/main/java/io/mosip/injivcrenderer/InjiVcRendererInterface.kt
@@ -1,0 +1,5 @@
+package io.mosip.injivcrenderer
+
+interface InjiVcRendererInterface {
+    fun replaceSVGTemplatePlaceholders(template: String, data: Map<String, Any>): String
+}
