--- kotlin/svgrenderer/build.gradle.kts
@@ -1,74 +1,0 @@
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.jetbrains.kotlin.android)
-    `maven-publish`
-}
-
-android {
-    namespace = "io.mosip.svgrenderer"
-    compileSdk = 34
-
-    defaultConfig {
-        minSdk = 24
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_21
-        targetCompatibility = JavaVersion.VERSION_21
-    }
-    kotlinOptions {
-        jvmTarget = "21"
-    }
-}
-
-dependencies {
-
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
-    implementation(libs.material)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.espresso.core)
-}
-
-
-tasks {
-    register<Wrapper>("wrapper") {
-        gradleVersion = "8.5"
-    }
-}
-
-tasks.register<Jar>("jarRelease") {
-    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-    dependsOn("assembleRelease")
-    from("build/intermediates/javac/release/classes") {
-        include("**/*.class")
-    }
-    from("build/tmp/kotlin-classes/release") {
-        include("**/*.class")
-    }
-    manifest {
-        attributes["Implementation-Title"] = project.name
-        attributes["Implementation-Version"] = "1.0-SNAPSHOT"
-    }
-    archiveBaseName.set("${project.name}-release")
-    archiveVersion.set("1.0-SNAPSHOT")
-    destinationDirectory.set(layout.buildDirectory.dir("libs"))
-}
-apply(from = "publish-artifact.gradle")
-tasks.register("generatePom") {
-    dependsOn("generatePomFileForAarPublication", "generatePomFileForJarReleasePublication")
-}
-
