--- kotlin/svgrenderer/src/main/java/io/mosip/svgrenderer/SvgRendererInterface.kt
@@ -1,5 +1,0 @@
-package io.mosip.svgrenderer
-
-interface SvgRendererInterface {
-    fun replaceSVGTemplatePlaceholders(template: String, data: Map<String, Any>): String
-}
